rename core repo -> label_repo
