+++ knowde/complex/definition/repo/definition.py
@@ -24,9 +24,9 @@ from knowde.complex.definition.repo.mark import (
     remark_sentence,
 )
 from knowde.complex.definition.repo.statistics import statistics_query
-from knowde.core.repo.base import RelBase
-from knowde.core.repo.query import QueryResult, query_cypher
-from knowde.core.repo.rel import RelUtil
+from knowde.core.label_repo.base import RelBase
+from knowde.core.label_repo.query import QueryResult, query_cypher
+from knowde.core.label_repo.rel import RelUtil
 from knowde.primitive.sentence import LSentence
 from knowde.primitive.sentence.domain import Sentence
 from knowde.primitive.term import LTerm, TermUtil
