+++ knowde/complex/person/repo/label.py
@@ -3,9 +3,9 @@ from neomodel import StringProperty, ZeroOrOne
 
 from knowde.complex.person.domain.person import Person
 from knowde.core.domain.domain import Entity
-from knowde.core.repo.base import LBase
-from knowde.core.repo.rel import RelUtil
-from knowde.core.repo.util import LabelUtil
+from knowde.core.label_repo.base import LBase
+from knowde.core.label_repo.rel import RelUtil
+from knowde.core.label_repo.util import LabelUtil
 from knowde.primitive.time.domain.period import Period
 from knowde.primitive.time.repo.label import LTime
