+++ knowde/feature/reference/repo/definition.py
@@ -12,7 +12,7 @@ from knowde.complex.definition.repo.definition import (
     q_stats_def,
 )
 from knowde.complex.definition.repo.label import REL_DEF_LABEL
-from knowde.core.repo.query import query_cypher
+from knowde.core.label_repo.query import query_cypher
 from knowde.feature.reference.domain import RefDefinition, RefDefinitions
 from knowde.feature.reference.dto import RefDefParam  # noqa: TCH001
 from knowde.primitive.reference.domain import Book
