+++ knowde/feature/reference/repo/rel.py
@@ -1,7 +1,7 @@
 """Rel Label."""
 from neomodel import DateProperty
 
-from knowde.core.repo.base import LBase
+from knowde.core.label_repo.base import LBase
 
 
 class RelWrite(LBase):
