+++ knowde/primitive/proposition/repo/label.py
@@ -1,8 +1,8 @@
 """neomodel label."""
 from neomodel import StringProperty
 
-from knowde.core.repo.base import LBase
-from knowde.core.repo.util import LabelUtil
+from knowde.core.label_repo.base import LBase
+from knowde.core.label_repo.util import LabelUtil
 from knowde.primitive.proposition.domain import Proposition
